moved all functions from operations file to methods of MPS
